bookclub_core/requirements.txt
pandas
openpyxl
xlrd
google-genai
python-dotenv
python-docx